Update documentation to reflect current state.
The documentation has been updated to reflect how the code has changed
to focus on the new hooking system and the ability to interact with the
host system. The old code has been moved to PyBeebMinimal.py as part
of this.
